tests-#12. Move each test next to actual implementation for easier navigation
